Moved PlayerInfo class in db manager
