+++ src/main/java/union/menus/AccountContext.java
@@ -5,8 +5,8 @@ import union.base.command.UserContextMenu;
 import union.base.command.UserContextMenuEvent;
 import union.objects.CmdAccessLevel;
 import union.objects.CmdModule;
-import union.objects.PlayerInfo;
 import union.objects.constants.Constants;
+import union.utils.database.managers.UnionPlayerManager.PlayerInfo;
 
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.entities.User;
